+++ speedwagon/__main__.py
@@ -1,6 +1,7 @@
 import logging
 import sys
 import speedwagon
+import speedwagon.startup
 import speedwagon.gui
 logger = logging.getLogger()
 logger.setLevel(logging.DEBUG)
@@ -8,13 +9,14 @@ logger.addHandler(logging.StreamHandler())
 
 
 def main():
-    print("{}: {}".format(speedwagon.__name__, speedwagon.__version__))
+    startup_settings = speedwagon.startup.parse_args()
 
     if len(sys.argv) > 1 and sys.argv[1] == "--pytest":
         import pytest  # type: ignore
         sys.exit(pytest.main(sys.argv[2:]))
     else:
-        speedwagon.gui.main()
+        print("{}: {}".format(speedwagon.__name__, speedwagon.__version__))
+        speedwagon.gui.main(startup_settings)
 
 
 if __name__ == '__main__':