+++ speedwagon/tools/tool_convertTifftoHathiTrustJP2.py
@@ -9,6 +9,7 @@ from speedwagon import worker
 from speedwagon.tools import options
 from speedwagon.job import AbsTool
 import pykdu_compress
+from py3exiv2bind.core import set_dpi
 
 
 class UserArgs(enum.Enum):
@@ -66,18 +67,34 @@ class ConvertFile(AbsProcessStrategy):
                                         basename + ".jp2"
                                         )
 
-        pykdu_compress.kdu_compress_cli(
-            "-i {} "
-            "Clevels=5 "
-            "Clayers=8 "
-            "Corder=RLCP "
-            "Cuse_sop=yes "
-            "Cuse_eph=yes "
-            "Cmodes=RESET|RESTART|CAUSAL|ERTERM|SEGMARK "
-            "-no_weights "
-            "-slope 42988 "
-            "-jp2_space sRGB "
-            "-o {}".format(source_file, output_file_path))
+        # pykdu_compress.kdu_compress_cli(
+        #     "-i {} "
+        #     "Clevels=5 "
+        #     "Clayers=8 "
+        #     "Corder=RLCP "
+        #     "Cuse_sop=yes "
+        #     "Cuse_eph=yes "
+        #     "Cmodes=RESET|RESTART|CAUSAL|ERTERM|SEGMARK "
+        #     "-no_weights "
+        #     "-slope 42988 "
+        #     "-jp2_space sRGB "
+        #     "-o {}".format(source_file, output_file_path))
+
+        in_args = [
+            "Clevels=5",
+            "Clayers=8",
+            "Corder=RLCP",
+            "Cuse_sop=yes",
+            "Cuse_eph=yes",
+            "Cmodes=RESET|RESTART|CAUSAL|ERTERM|SEGMARK",
+            "-no_weights",
+            "-slope", "42988",
+            "-jp2_space", "sRGB",
+        ]
+        pykdu_compress.kdu_compress_cli2(
+            source_file, output_file_path, in_args=in_args
+        )
+        set_dpi(output_file_path, x=400, y=400)
 
         self.status = "Generated {}".format(output_file_path)